Split test_packages_and_services.py into two files
